1. New cache-out algorithm PORE+. 2. Re-Design FIFO function implement in SMR Simulator.
