+++ src/Makefile
@@ -9,7 +9,8 @@ CPPFLAGS += -I$(SMR_SSD_CACHE_DIR) -I$(SMR_SSD_CACHE_DIR)/smr-simulator -I$(SMR_
 
 RM = rm -rf
 RMSHM = rm -f /dev/shm/*
-OBJS = global.o report.o timerUtils.o shmlib.o hashtable_utils.o cache.o trace2call.o main.o losertree4pore.o pore.o lru.o lru_private.o lru_batch.o smr-simulator.o inner_ssd_buf_table.o
+OBJS = global.o report.o timerUtils.o shmlib.o hashtable_utils.o cache.o trace2call.o main.o pore_plus.o losertree4pore.o pore.o lru.o lru_private.o lru_batch.o smr-simulator.o inner_ssd_buf_table.o simulator_logfifo.o
+
 
 all: $(OBJS) smr-ssd-cache
 	@echo 'Successfully built smr-ssd-cache...'
@@ -41,6 +42,9 @@ trace2call.o: trace2call.c
 main.o: main.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c $?
 
+pore_plus.o: strategy/pore_plus.c
+	$(CC) $(CPPFLAGS) $(CFLAGS) -c $?
+
 losertree4pore.o: strategy/losertree4pore.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c $?
 
@@ -63,6 +67,9 @@ smr-simulator.o: smr-simulator/smr-simulator.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c $?
 
 inner_ssd_buf_table.o: smr-simulator/inner_ssd_buf_table.c
+	$(CC) $(CPPFLAGS) $(CFLAGS) -c $?
+
+simulator_logfifo.o: smr-simulator/simulator_logfifo.c
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c $?
 
 clean:
