+++ src/global.c
@@ -23,11 +23,12 @@ blksize_t NBLOCK_SMR_FIFO;
 blksize_t NBANDTables = 2621952;
 blksize_t SSD_SIZE = 4096;
 const blksize_t BLCKSZ = 4096;
-const blksize_t NZONES = 194180;    // NZONES * ZONESZ =
+const blkcnt_t  NZONES = 194180;    // NZONES * ZONESZ =
 const blksize_t ZONESZ = 18874368;//18MB    // Unit: Byte.
-blksize_t WRITEAMPLIFICATION = 100;
-blksize_t NCOLDBAND = 1;
-char smr_device[] = "/dev/sda";
+
+char simu_smr_fifo_device[] = "/dev/sdb1";
+char simu_smr_smr_device[] = "/dev/sdb2";
+char smr_device[] = "/dev/sdb";
 char ssd_device[] = "/mnt/ssd/ssd";
 char ram_device[1024];
