+++ src/strategy/pore_plus.h
